# Use Reflection.Emit to generate IPerformanceCounterSet implementation 