--- a/.gitignore
+++ b/.gitignore
@@ -28,6 +28,10 @@ local.properties
 # PDT-specific
 .buildpath
 
+#################
+## IDEA
+#################
+.idea
 
 #################
 ## Visual Studio