--- a/.gitignore
+++ b/.gitignore
@@ -44,6 +44,7 @@ local.properties
 *.suo
 *.user
 *.sln.docstates
+.vs/
 
 # Build results
 [Dd]ebug/