--- a/NuGet.Pack.cmd
+++ b/NuGet.Pack.cmd
@@ -1,7 +1,0 @@
-@echo off
-
-set NUGET="src/.nuget/nuget.exe"
-
-if not exist output ( mkdir output )
-
-%NUGET% Pack "src/PerformanceCounters/PerformanceCounters.nuspec" -OutputDirectory output