--- a/src/PerformanceCounters.WebApi/packages.config
+++ b/src/PerformanceCounters.WebApi/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.AspNet.WebApi.Client" version="4.0.20710.0" targetFramework="net451" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="4.0.20710.0" targetFramework="net451" />
-  <package id="Microsoft.Net.Http" version="2.0.20710.0" targetFramework="net451" />
-  <package id="Newtonsoft.Json" version="4.5.6" targetFramework="net451" />
-</packages>