--- a/src/PerformanceCounters.sln
+++ b/src/PerformanceCounters.sln
@@ -5,19 +5,12 @@ VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PerformanceCounters", "PerformanceCounters\PerformanceCounters.csproj", "{66E68F50-C73A-4C0A-9A51-AC68432EE715}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{4AA31A97-ABC7-49A7-9594-289F5EA45554}"
-	ProjectSection(SolutionItems) = preProject
-		.nuget\NuGet.Config = .nuget\NuGet.Config
-		.nuget\NuGet.exe = .nuget\NuGet.exe
-		.nuget\NuGet.targets = .nuget\NuGet.targets
-	EndProjectSection
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleApplication", "SampleApplication\SampleApplication.csproj", "{0DA52B2A-B454-456E-BB4D-8B1F1AEB7A0C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{93BB8265-A93D-4F30-BD65-299F90B84CDB}"
-	ProjectSection(SolutionItems) = preProject
-		ReadMe.md = ReadMe.md
-	EndProjectSection
+ProjectSection(SolutionItems) = preProject
+	..\readme.markdown = ..\readme.markdown
+EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PerformanceCounters.WebApi", "PerformanceCounters.WebApi\PerformanceCounters.WebApi.csproj", "{1E6E5679-B21F-45A3-86D4-952C80EB181A}"
 EndProject