--- a/src/PerformanceCounters/IPerformanceCounter.cs
+++ b/src/PerformanceCounters/IPerformanceCounter.cs
@@ -4,10 +4,6 @@ namespace NeedfulThings.PerformanceCounters
 
 	public interface IPerformanceCounter : IReadOnlyPerformanceCounter
 	{
-		string CounterName { get; }
-
-		PerformanceCounterType CounterType { get; }
-
 		void Increment();
 
 		void IncrementBy(long value);