--- a/src/PerformanceCounters/IPerformanceCounterSet.cs
+++ b/src/PerformanceCounters/IPerformanceCounterSet.cs
@@ -6,6 +6,6 @@
 	{
 		string CategoryName { get; }
 
-		IReadOnlyCollection<IPerformanceCounter> Counters { get; }
+		IReadOnlyCollection<IReadOnlyPerformanceCounter> Counters { get; }
 	}
 }