--- a/src/PerformanceCounters/IReadOnlyPerformanceCounter.cs
+++ b/src/PerformanceCounters/IReadOnlyPerformanceCounter.cs
@@ -1,9 +1,15 @@
+using System.Diagnostics;
+
 namespace NeedfulThings.PerformanceCounters
 {
 	using System;
 
 	public interface IReadOnlyPerformanceCounter : IDisposable
 	{
+		string CounterName { get; }
+
+		PerformanceCounterType CounterType { get; }
+		
 		float NextValue();
 	}
 }