--- a/src/PerformanceCounters/PerformanceCounterFactory.cs
+++ b/src/PerformanceCounters/PerformanceCounterFactory.cs
@@ -1,14 +1,14 @@
+using System.Linq;
+using System.Reflection;
+
 namespace NeedfulThings.PerformanceCounters
 {
 	using System;
 	using System.Collections.Generic;
 	using System.Diagnostics;
-	using Castle.DynamicProxy;
 
 	public static class PerformanceCounterFactory
 	{
-		private static readonly ProxyGenerator ProxyGenerator = new ProxyGenerator();
-
 		private static readonly Dictionary<Type, IPerformanceCounterSet> _counters =
 			new Dictionary<Type, IPerformanceCounterSet>();
 
@@ -47,7 +47,20 @@ namespace NeedfulThings.PerformanceCounters
 				IPerformanceCounterSet counterSet;
 				if (!_counters.TryGetValue(typeof (T), out counterSet))
 				{
-					counterSet = ProxyGenerator.CreateInterfaceProxyWithoutTarget<T>(new PerformanceCountersInterceptor<T>());
+					var counterSetType = CounterSetTypeEmitter.GeneratePerformanceCounterSetImplementation(typeof(T));
+					var category = Helper.GetCategoryAttribute(typeof(T));
+					var performanceCounters = GetPerformanceCounters<T>();
+
+					var arguments = new object[performanceCounters.Count + 2];
+					arguments[0]= category.CategoryName;
+					arguments[1]= performanceCounters;
+					for (int i = 0; i < performanceCounters.Count; i++)
+					{
+						arguments[i + 2] = performanceCounters[i];
+					}
+
+					counterSet = (T)Activator.CreateInstance(counterSetType, arguments.ToArray());
+
 					_counters.Add(typeof (T), counterSet);
 				}
 
@@ -55,7 +68,41 @@ namespace NeedfulThings.PerformanceCounters
 			}
 		}
 
-		internal static IPerformanceCounter GetInstance(PerformanceCounterCategoryAttribute categoryAttribute,
+		private static IReadOnlyList<IReadOnlyPerformanceCounter> GetPerformanceCounters<T>()
+		{
+			var type = typeof (T);
+			var categoryAttribute = Helper.GetCategoryAttribute(type);
+			if (categoryAttribute == null)
+			{
+				throw new ArgumentException();
+			}
+
+			var counters = new List<IReadOnlyPerformanceCounter>();
+			foreach (var propertyInfo in type.GetProperties())
+			{
+				var counterAttribute = GetCounterAttribute(propertyInfo);
+				if (counterAttribute == null)
+				{
+					continue;
+				}
+
+				var getMethod = propertyInfo.GetGetMethod();
+				if (getMethod == null)
+				{
+					throw new InvalidProgramException();
+				}
+
+				var counter = GetInstance(categoryAttribute, counterAttribute,
+					propertyInfo.PropertyType ==
+					typeof (IReadOnlyPerformanceCounter));
+
+				counters.Add(counter);
+			}
+
+			return counters;
+		}
+
+		private static IPerformanceCounter GetInstance(PerformanceCounterCategoryAttribute categoryAttribute,
 		                                                PerformanceCounterAttribute counterAttribute, bool readOnly)
 		{
 			var categoryName = categoryAttribute.CategoryName;
@@ -81,6 +128,15 @@ namespace NeedfulThings.PerformanceCounters
 			}
 
 			return new NullPerformanceCounter(counterName, counterType);
+		}
+
+		private static PerformanceCounterAttribute GetCounterAttribute(PropertyInfo propertyInfo)
+		{
+			var attribute =
+				(PerformanceCounterAttribute)
+				propertyInfo.GetCustomAttributes(typeof (PerformanceCounterAttribute), false).FirstOrDefault();
+
+			return attribute;
 		}
 	}
 }