--- a/src/PerformanceCounters/PerformanceCounters.csproj
+++ b/src/PerformanceCounters/PerformanceCounters.csproj
@@ -32,16 +32,13 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="Castle.Core">
-      <HintPath>..\packages\Castle.Core.3.2.0\lib\net45\Castle.Core.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Configuration.Install" />
     <Reference Include="System.Core" />
     <Reference Include="Microsoft.CSharp" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CounterSetTypeEmitter.cs" />
     <Compile Include="Helper.cs" />
     <Compile Include="IPerformanceCounter.cs" />
     <Compile Include="IPerformanceCounterSet.cs" />
@@ -51,11 +48,7 @@
     <Compile Include="PerformanceCounterCategoryAttribute.cs" />
     <Compile Include="PerformanceCounterFactory.cs" />
     <Compile Include="PerformanceCounterProxy.cs" />
-    <Compile Include="PerformanceCountersInterceptor.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="packages.config" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\nuget.targets" />