--- a/src/PerformanceCounters/packages.config
+++ b/src/PerformanceCounters/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<packages>
-  <package id="Castle.Core" version="3.2.0" targetFramework="net45" />
-</packages>