… src/.nuget/NuGet.exe
… src/.nuget/NuGet.targets
… src/PerformanceCounters.WebApi/Properties/AssemblyInfo.cs
… src/PerformanceCounters/PerformanceCountersInterceptor.cs
… src/PerformanceCounters/Properties/AssemblyInfo.cs
… src/PerformanceCounters/install.ps1
… src/SampleApplication/Properties/AssemblyInfo.cs
… src/SampleApplication/SampleApplication.csproj
… src/SampleApplication/packages.config