Improve performance and avoid memory allocation on accessing `IPerformanceCounter`.

Now:

|                      Method |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------------------------- |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
| IncrementPerformanceCounter | 10.871 ns | 0.2447 ns | 0.4887 ns |           - |           - |           - |                   - |
| IncrementPurelPerfCounter |  8.135 ns | 0.1556 ns | 0.1455 ns |           - |           - |           - |                   - |
| IncrementPreviousPerformanceCounter | 102.371 ns | 1.445 ns | 4.873 ns |           - |           - |           - |                   106B |